Functionality extracted to its own packages.
