+++ pkg/attributesdecorator/go.mod
@@ -1,0 +1,16 @@
+module github.com/solarwinds/solarwinds-otel-collector-contrib/pkg/attributesdecorator
+
+go 1.24.3
+
+require go.opentelemetry.io/collector/pdata v1.33.0
+
+require (
+	github.com/gogo/protobuf v1.3.2 // indirect
+	go.uber.org/multierr v1.11.0 // indirect
+	golang.org/x/net v0.39.0 // indirect
+	golang.org/x/sys v0.32.0 // indirect
+	golang.org/x/text v0.24.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20250218202821-56aae31c358a // indirect
+	google.golang.org/grpc v1.72.2 // indirect
+	google.golang.org/protobuf v1.36.6 // indirect
+)
